bot/requirements.txt
python-dotenv
requests
pandas
numpy
matplotlib
python-telegram-bot
python-telegram-bot[job-queue]
APScheduler==3.10.1